feat: coupon 추가 구현

+++ backend/sin/src/main/java/sin/sin/domain/coupon/Coupon.java
@@ -32,7 +32,7 @@ public class Coupon {
     private Timestamp expirationDate;
 
     @Column(nullable = false)
-    private float discountPercent;
+    private int discountPercent;
 
     @Column(nullable = false)
     private int discountAmount;
